requirements.txt
alabaster==1.0.0
asgiref==3.8.1
babel==2.16.0
certifi==2022.6.15
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.6.0
distlib==0.3.5
Django==5.1.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
docutils==0.21.2
filelock==3.4.2
groundwork-sphinx-theme==1.1.1
idna==3.8
imagesize==1.4.1
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
platformdirs==2.6.2
Pygments==2.18.0
pytz==2024.1
requests==2.32.3
six==1.16.0
snowballstemmer==2.2.0
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==2.2.2
virtualenv==20.15.1
virtualenv-clone==0.5.7
